+++ engine/src/search_engine/tree/node.rs
@@ -18,7 +18,7 @@ pub struct Node {
     mv: AtomicU16,
     visit_count: AtomicU32,
     cumulative_score: AtomicWDLScore,
-    base_score: AtomicU32,
+    base_score: AtomicU64,
     squared_score: AtomicU64,
     children_start_index: IndexLock,
     children_count: AtomicU8,
@@ -34,7 +34,7 @@ impl Clone for Node {
             mv: AtomicU16::new(self.mv.load(Ordering::Relaxed)),
             visit_count: AtomicU32::new(self.visit_count.load(Ordering::Relaxed)),
             cumulative_score: self.cumulative_score.clone(),
-            base_score: AtomicU32::new(self.base_score.load(Ordering::Relaxed)),
+            base_score: AtomicU64::new(self.base_score.load(Ordering::Relaxed)),
             squared_score: AtomicU64::new(self.squared_score.load(Ordering::Relaxed)),
             children_start_index: self.children_start_index.clone(),
             children_count: AtomicU8::new(self.children_count.load(Ordering::Relaxed)),
@@ -52,7 +52,7 @@ impl Node {
             mv: AtomicU16::new(0),
             visit_count: AtomicU32::new(0),
             cumulative_score: AtomicWDLScore::default(),
-            base_score: AtomicU32::new(0),
+            base_score: AtomicU64::new(0),
             squared_score: AtomicU64::new(0),
             children_start_index: IndexLock::new(NodeIndex::NULL),
             children_count: AtomicU8::new(0),
@@ -111,8 +111,8 @@ impl Node {
     }
 
     #[inline]
-    pub fn base_score(&self) -> f32 {
-        f32::from_bits(self.base_score.load(Ordering::Relaxed))
+    pub fn base_score(&self) -> f64 {
+        f64::from_bits(self.base_score.load(Ordering::Relaxed))
     }
 
     #[inline]
@@ -162,7 +162,7 @@ impl Node {
 
     #[inline]
     pub fn set_base_score(&self, score: WDLScore) {
-        self.base_score.store((score.single() as f32).to_bits(), Ordering::Relaxed);
+        self.base_score.store((score.single()).to_bits(), Ordering::Relaxed);
     }
 
     #[inline]
